refactor code movie-service
