+++ services/movie-service/prisma/schema.prisma
@@ -26,6 +26,7 @@ model Movie {
   trailerUrl        String
   trailerPublicId   String
   genres            Genres[]
+  isActive          Boolean  @default(true)
   createdAt         DateTime @default(now())
   updatedAt         DateTime @updatedAt
