+++ services/movie-service/src/routes/movie.routes.ts
@@ -5,16 +5,33 @@ import { upload } from "../middlewares/upload.js";
 
 const router = express.Router();
 
-router.get("/genres", MovieController.getGenres);
 router.get("/", MovieController.getMovies);
 router.get("/:movieId", MovieController.getMovieById);
-router.post("/", verifyToken, authorizeRole("ADMIN"), upload.fields([{ name: "thumbnail" }, { name: "trailer" }]), MovieController.createMovie);
-router.put("/:movieId", verifyToken, authorizeRole("ADMIN"), upload.fields([{ name: "thumbnail" }, { name: "trailer" }]), MovieController.updateMovieById);
-router.delete("/:movieId", verifyToken, authorizeRole("ADMIN"), MovieController.deleteMovieById);
-router.get("/genres/:genreId", MovieController.getGenreById);
-router.post("/genres", verifyToken, authorizeRole("ADMIN"), MovieController.createGenre);
-router.put("/genres/:genreId", verifyToken, authorizeRole("ADMIN"), MovieController.updateGenreById);
-router.put("/genres/archive/:genreId", verifyToken, authorizeRole("ADMIN"), MovieController.archiveGenreById);
-router.put("/genres/restore/:genreId", verifyToken, authorizeRole("ADMIN"), MovieController.restoreGenreById);
+router.post(
+  "/",
+  verifyToken,
+  authorizeRole("ADMIN"),
+  upload.fields([{ name: "thumbnail" }, { name: "trailer" }]),
+  MovieController.createMovie
+);
+router.put(
+  "/:movieId",
+  verifyToken,
+  authorizeRole("ADMIN"),
+  upload.fields([{ name: "thumbnail" }, { name: "trailer" }]),
+  MovieController.updateMovieById
+);
+router.put(
+  "/archive/:movieId",
+  verifyToken,
+  authorizeRole("ADMIN"),
+  MovieController.archiveMovieById
+);
+router.put(
+  "/restore/:movieId",
+  verifyToken,
+  authorizeRole("ADMIN"),
+  MovieController.restoreMovieById
+);
 
 export default router;
